Replace Express API with Netlify functions
