+++ package.json
@@ -3,15 +3,11 @@
   "version": "3.0.0",
   "private": true,
   "description": "Advanced IELTS Speaking Test Evaluator with pronunciation analysis and stress pattern detection",
-  "main": "server.js",
   "scripts": {
-    "start": "node server.js",
-    "dev": "concurrently \"nodemon server.js\" \"npm run client\"",
-    "client": "react-scripts start",
+    "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "heroku-postbuild": "npm run build"
+    "eject": "react-scripts eject"
   },
   "dependencies": {
     "axios": "^1.6.0",
